feat: remove awsDataTransferCostTrend widget and update baseTrend widget
